+++ tests/Neo.SmartContract.Framework.UnitTests/ManifestExtraTest.cs
@@ -1,0 +1,18 @@
+using Microsoft.VisualStudio.TestTools.UnitTesting;
+using Neo.SmartContract.Testing;
+
+namespace Neo.SmartContract.Framework.UnitTests.Old
+{
+    [TestClass]
+    public class ManifestExtraTest
+    {
+        [TestMethod]
+        public void TestExtraAttribute()
+        {
+            var extra = Contract_ExtraAttribute.Manifest.Extra;
+
+            Assert.AreEqual("Neo", extra["Author"]?.GetString());
+            Assert.AreEqual("dev@neo.org", extra["E-mail"]?.GetString());
+        }
+    }
+}
